requirements_pip.txt
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
filelock==3.12.2
fonttools==4.41.1
gmpy2==2.1.2
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
Pillow==9.4.0
pip==23.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
secml==0.15.6
setuptools==68.0.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
wheel==0.40.0
zipp==3.16.2
